Use before and after lifecycles to sync children
